# backend/openapi_new.json
{"openapi":"3.1.0","info":{"title":"BharatAI API","description":"Core API for Sovereign Gov AI Platform","version":"1.0.0"},"paths":{"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/token":{"post":{"tags":["auth"],"summary":"Login Swagger","operationId":"login_swagger_api_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_swagger_api_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Read Users Me","operationId":"read_users_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/chat/message":{"post":{"tags":["chat"],"summary":"Chat Message","operationId":"chat_message_api_chat_message_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_chat_message_api_chat_message_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/chat/conversations":{"get":{"tags":["chat"],"summary":"Get Conversations","operationId":"get_conversations_api_chat_conversations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/chat/conversations/{conv_id}/messages":{"get":{"tags":["chat"],"summary":"Get Messages","operationId":"get_messages_api_chat_conversations__conv_id__messages_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"conv_id","in":"path","required":true,"schema":{"type":"integer","title":"Conv Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/audit/logs":{"get":{"tags":["audit"],"summary":"Get Audit Logs","operationId":"get_audit_logs_api_audit_logs_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ministry","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ministry"}},{"name":"flagged_only","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Flagged Only"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/audit/stats":{"get":{"tags":["audit"],"summary":"Get Stats","operationId":"get_stats_api_audit_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/audit/export":{"get":{"tags":["audit"],"summary":"Export Audit Csv","operationId":"export_audit_csv_api_audit_export_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/documents/upload":{"post":{"tags":["documents"],"summary":"Upload Document","operationId":"upload_document_api_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_documents_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/documents/list":{"get":{"tags":["documents"],"summary":"List Documents","operationId":"list_documents_api_documents_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/models":{"get":{"tags":["models"],"summary":"List Models","description":"Proxy GET /api/tags from Ollama and return a flat list of model names.\nAny logged-in user can call this endpoint.","operationId":"list_models_api_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/rag/debug":{"post":{"tags":["RAG Debug"],"summary":"Debug Rag","description":"Debug endpoint — shows exactly what RAG retrieves.\nAdmin only.","operationId":"debug_rag_api_rag_debug_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DebugQueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/rag/stats/{ministry}":{"get":{"tags":["RAG Debug"],"summary":"Ministry Stats","description":"Get knowledge base stats for a ministry","operationId":"ministry_stats_api_rag_stats__ministry__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ministry","in":"path","required":true,"schema":{"type":"string","title":"Ministry"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/policy/rules":{"get":{"tags":["Policy Engine"],"summary":"List Rules","description":"List all policy rules with trigger counts","operationId":"list_rules_api_policy_rules_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"ministry","in":"query","required":false,"schema":{"type":"string","title":"Ministry"}},{"name":"active_only","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Active Only"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Policy Engine"],"summary":"Create Rule","description":"Create a new policy rule","operationId":"create_rule_api_policy_rules_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRuleRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/policy/rules/{rule_id}":{"patch":{"tags":["Policy Engine"],"summary":"Update Rule","description":"Update or toggle a policy rule","operationId":"update_rule_api_policy_rules__rule_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"integer","title":"Rule Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRuleRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Policy Engine"],"summary":"Delete Rule","description":"Soft delete — deactivates rule","operationId":"delete_rule_api_policy_rules__rule_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"integer","title":"Rule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/policy/rules/stats":{"get":{"tags":["Policy Engine"],"summary":"Policy Stats","description":"Get policy engine statistics","operationId":"policy_stats_api_policy_rules_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/policy/rules/test":{"post":{"tags":["Policy Engine"],"summary":"Test Rule","description":"Test endpoint — runs a query through policy engine\nwithout actually sending to LLM.\nPerfect for admins to test new rules before activating.","operationId":"test_rule_api_policy_rules_test_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}},{"name":"ministry","in":"query","required":false,"schema":{"type":"string","default":"General","title":"Ministry"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_chat_message_api_chat_message_post":{"properties":{"conversation_id":{"type":"integer","title":"Conversation Id"},"message":{"type":"string","title":"Message"},"ministry_context":{"type":"string","title":"Ministry Context","default":"General"},"language":{"type":"string","title":"Language","default":"English"},"model":{"type":"string","title":"Model"}},"type":"object","required":["message"],"title":"Body_chat_message_api_chat_message_post"},"Body_login_swagger_api_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_swagger_api_auth_token_post"},"Body_upload_document_api_documents_upload_post":{"properties":{"ministry":{"type":"string","title":"Ministry"},"doc_type":{"type":"string","title":"Doc Type","default":"general"},"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["ministry","file"],"title":"Body_upload_document_api_documents_upload_post"},"CreateRuleRequest":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","default":""},"ministry":{"type":"string","title":"Ministry","default":"ALL"},"pattern":{"type":"string","title":"Pattern"},"pattern_type":{"type":"string","title":"Pattern Type","default":"keyword"},"action":{"type":"string","title":"Action"},"clearance_required":{"type":"integer","title":"Clearance Required","default":0}},"type":"object","required":["name","pattern","action"],"title":"CreateRuleRequest"},"DebugQueryRequest":{"properties":{"query":{"type":"string","title":"Query"},"ministry":{"type":"string","title":"Ministry"}},"type":"object","required":["query","ministry"],"title":"DebugQueryRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"UpdateRuleRequest":{"properties":{"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Action"},"clearance_required":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Clearance Required"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","title":"UpdateRuleRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/token"}}}}}}